requirements.txt
asgiref==3.5.2
Django==4.1.3
python-decouple==3.6
sqlparse==0.4.3
tzdata==2022.6

pip~=21.3.1
wheel~=0.37.1
Pillow~=9.3.0
setuptools~=60.2.0
python-decouple~=3.6